ev_rental_project/requirements.txt
Django>=5.0,<6.0
psycopg2-binary
dspy-ai
google-generativeai
python-dotenv
